# backend/swagger-spec.json
{"openapi":"3.0.0","paths":{"/subtitleToken/{id}":{"get":{"operationId":"AppController_getTokenizedSubtitles","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/TokenizedSubtitleResponseDto"}}}}}}}},"/subtitleToken":{"get":{"operationId":"AppController_getTokenizedSubtitle","parameters":[{"name":"text","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenizedSubtitleResponseDto"}}}}}}},"/subtitles/{id}":{"get":{"operationId":"AppController_getSubtitles","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SubtitleResponseDto"}}}}}}}}},"info":{"title":"Magni Miner","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"TokenizedSubtitleResponseDto":{"type":"object","properties":{"tokens":{"type":"array","items":{"type":"object","properties":{"reading":{"type":"string"},"text":{"type":"string"},"kana":{"type":"string"},"score":{"type":"number"},"seq":{"type":"number"},"compound":{"type":"array","items":{"type":"string"}},"components":{"type":"array","items":{"type":"object","properties":{"reading":{"type":"string"},"text":{"type":"string"},"kana":{"type":"string"},"score":{"type":"number"},"seq":{"type":"number"},"conj":{"type":"object","properties":{"prop":{"type":"array","items":{"type":"object","properties":{"pos":{"type":"string"},"type":{"type":"string"}},"required":["pos","type"]}},"reading":{"type":"string"},"gloss":{"type":"object","properties":{"pos":{"type":"string"},"gloss":{"type":"string"},"info":{"type":"string"}},"required":["pos","gloss","info"]}},"required":["prop","reading","gloss"]}},"required":["reading","text","kana","score","seq","conj"]}},"gloss":{"type":"array","items":{"type":"object","properties":{"pos":{"type":"string"},"gloss":{"type":"string"},"info":{"type":"string"}},"required":["pos","gloss","info"]}},"conj":{"type":"array","items":{"type":"object","properties":{"prop":{"type":"array","items":{"type":"object","properties":{"pos":{"type":"string"},"type":{"type":"string"}},"required":["pos","type"]}},"reading":{"type":"string"},"gloss":{"type":"object","properties":{"pos":{"type":"string"},"gloss":{"type":"string"},"info":{"type":"string"}},"required":["pos","gloss","info"]}},"required":["prop","reading","gloss"]}}},"required":["reading","text","kana","score","seq","compound","components","gloss","conj"]}},"start":{"type":"number"},"end":{"type":"number"}},"required":["tokens","start","end"]},"SubtitleResponseDto":{"type":"object","properties":{"text":{"type":"string"},"start":{"type":"number"},"end":{"type":"number"}},"required":["text","start","end"]}}}}